added end to end tests with chai
